requirements.txt
fastapi
uvicorn
pymongo
requests
python-jose[cryptography]
passlib
bcrypt==4.0.1
openai
python-multipart
pydantic
python-dotenv
 
